Refactored client, server, and etc modules to more cleanly separate client/server functionality from etc-related calculations. The separation is not complete, but the etc-dependency is now encapsulated in one method of one class in server.py.
The server.py module replaces the etcserver.py module, and should now be used to start the server. The etc_test.py module is still used in the same way to exercise the client/server communication with a few typical etc calls. 

Also implemented the ability to specify a port by means of the PYSYN_PORT environment variable. If this variable is not set, it will use the default port (specified in server.py and imported by client.py). This resolves ticket #47.


git-svn-id: https://aeon.stsci.edu/ssb/svn/astrolib/trunk/pysynphot@219 90a0a646-be8a-0410-bb88-9290da87bc01