Ticket #51: don't insert a default directory path if none is included in the filespec. Modified spparser.py/_handleIRAFName accordingly. (This should be refactored so we don't have both _handleIRAFName and irafconvert.)
Refactored locations.irafconvert into a more pythonic idiom; this included removing the '$' from the end of the keys in the convertdic.

Refactored spparser._handleThroughputFiles: Instead of checking a special lookup table in locations.py (now removed), try to return a spectrum and  if that fails, return a spectral element.

Deleted v0p2_test.py/ETCTestCase_Imag2.test8() on Donald's assurance
that this is an obsolete test case that no longer reflects current
ETC behavior.

Closes #51.


git-svn-id: https://aeon.stsci.edu/ssb/svn/astrolib/trunk/pysynphot@231 90a0a646-be8a-0410-bb88-9290da87bc01