Ticket #56: Raise exceptions if either the obsmode string or graph table is not well-formed. Specifically, if the obsmode is incomplete, or if there are unused keywords, or if the graphtable shows multiple matches for the given obsmode.
Ticket #57: Allow strings such as "band(johnson, v)" to go through the graph table successfully. Eliminated the Magnitude helper class; moved its methods to the SpectralElement class and changed the calling sequence of spectrum.setMagnitude to accept a SpectralElement and a scalar value.

As of this changeset, the tests in v0p2_test.py no longer pass when used with the real CDBS. If a special version of CDBS is used that replaces version 004 of the bessell h and johnson v files with version 003, then all but 5 tests pass: these five are failing with an unused keyword exception because parameterized keywords are not handled properly (see ticket #21). v0p2_test.py will be abandoned at this point and replaced by new test files.


git-svn-id: https://aeon.stsci.edu/ssb/svn/astrolib/trunk/pysynphot@248 90a0a646-be8a-0410-bb88-9290da87bc01