This is a refactoring of binning calculations in pysynphot.
Now pysynphot and other things can use the same code base for calculations
related to binning. This work was to be the foundation of #240 but other
discoveries along the way have expanded the required work on pysynphot, work
that will have to wait. So this is just getting in the foundational changes
that don't change anything.



git-svn-id: https://aeon.stsci.edu/ssb/svn/astrolib/trunk/pysynphot@3138 90a0a646-be8a-0410-bb88-9290da87bc01